requirements.txt
tensorflow>=2.15.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0  # Required for image handling 